review: harden iso9660 + cooked reader against malformed images
Fixes raised by Gemini, Copilot, and CodeRabbit on PR #7:

cooked: do remaining-bytes comparison in u64 rather than narrowing to
usize before the min. On 32-bit pointer-width targets like wasm32-wasi
the previous cast would truncate when total_cooked - logical_pos
exceeded 2^32 and return a premature EOF on >4 GB images.

iso9660: cap per-directory extent allocation at 16 MiB and refuse
nesting beyond 64 levels so a crafted image can't request a multi-GB
Vec or stack-overflow the recursive walker. Track visited directory
LBAs and error on revisits to defend against extents that loop back
to an ancestor under a different name.

extract: reject pre-existing symlinks at each child path during the
write walk. Without this, a populated output dir (or one used with
--force) could redirect create_dir_all/File::create outside the
extraction root.

README: tag the extract example block as bash to satisfy MD040.

Re-ran the end-to-end against PMagic_8: 1278 files, byte-for-byte
identical to the reference ISO mount.

+++ src/cooked.rs
@@ -129,8 +129,15 @@ impl<R: Read + Seek> Read for CookedSectorReader<R> {
         let user_start = self.user_data.start + within;
         let user_end = self.user_data.end;
         let available_in_sector = user_end - user_start;
-        let remaining_total = (total_cooked - self.logical_pos) as usize;
-        let n = min(buf.len(), min(available_in_sector, remaining_total));
+        // Do the comparison in u64 so the remaining-bytes count doesn't
+        // truncate on 32-bit pointer-width targets like wasm32-wasi for
+        // images larger than 4 GB. The final value is bounded by
+        // available_in_sector (<= 2048), so the usize cast is safe.
+        let remaining_total = total_cooked - self.logical_pos;
+        let n = min(
+            buf.len() as u64,
+            min(available_in_sector as u64, remaining_total),
+        ) as usize;
 
         buf[..n].copy_from_slice(&self.cache[user_start..user_start + n]);
         self.logical_pos += n as u64;
